pyproject.toml
[project]
name = "one-fraud"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agentops>=0.4.18",
    "aiohttp>=3.12.14",
    "anthropic>=0.59.0",
    "azure-identity>=1.23.1",
    "cohere>=5.16.1",
    "cryptography>=45.0.5",
    "faiss-cpu>=1.11.0.post1",
    "google-adk==1.3.0",
    "google-ai-generativelanguage==0.6.15",
    "google-api-core==2.25.1",
    "google-auth>=2.40.3",
    "google-generativeai==0.8.5",
    "groq>=0.30.0",
    "httpx>=0.28.1",
    "ipykernel>=6.30.0",
    "jupyter>=1.1.1",
    "litellm>=1.74.8",
    "mistralai>=1.9.3",
    "numpy>=1.21.0,<2.0.0",
    "ollama>=0.5.1",
    "openai>=1.97.1",
    "pandas>=2.3.1",
    "pre-commit>=4.2.0",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "rich>=14.1.0",
    "sentence-transformers>=5.0.0",
    "streamlit>=1.32.0",
    "tiktoken>=0.9.0",
    "together>=1.5.21",
    "typing-extensions>=4.14.1",
]
